✨ feat: se conectó la pasarela de pago con payload y la BD, también con la thnank you page
