+++ src/app/api/prueba2.js
@@ -1,3 +1,6 @@
+// Esta es una API que crea una preferencia de MercadoPago con los productos recibidos en el body
+const payload = require("payload");
+const getPayloadClient  = require("../../getPayload");
 const { MercadoPagoConfig, Preference } = require("mercadopago");
 
 const client = new MercadoPagoConfig({
@@ -14,9 +17,13 @@ const isProductValid = (product) => {
 
 const createPreference = async (req, res) => {
     try {
-        console.log(req.body);
+        const productsIDS = req.body.map((product) => product.id)
+        const userID = req.body[0].user_id;
+        const products = req.body.map(({ id, user_id, ...rest }) => rest);
 
-        const products = req.body;
+        console.log("esto es productosIDS: ",productsIDS);
+        console.log("esto es userID: ",userID);
+
 
         if (!Array.isArray(products) || products.length === 0) {
             return res.status(400).json({ error: "No se recibieron productos" });
@@ -27,22 +34,96 @@ const createPreference = async (req, res) => {
                 return res.status(400).json({ error: "Formato de producto inválido" });
             }
         }
-
+        console.log("Productos validos");
         const preference2 = new Preference(client);
+        
+                //#region ~~~~~~~~~~~~~~~~~~~~~~~~~~~ Payload truco ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+        // traer user de la DB
+        const { docs: users } = await payload.find({
+                    collection: 'users',
+                    where: {
+                        id: { equals: userID },
+                        },
+                    });
+        const user = users[0];
+
+        console.log("Usuario encontrado: ", user.id);
+
+        
+        // Crear la orden en la base de datos
+        const chizuOrder = await payload.create({
+            collection: 'orders',
+            data: {
+                preferenceId: null,
+                paymentId: null,
+                _isPaid: false,
+                products: productsIDS,
+                user: userID,
+                total: products.reduce((acc, product) => {
+                    return acc + product.unit_price;
+                }, 0),
+            },
+        });
+
+        console.log("Orden Creada payload: ", chizuOrder.id);
+        
+        await payload.update({
+            collection : 'users', 
+            id : userID, 
+            data : { 
+                ordenes : ( (user.ordenes_hist ?? []).length + 1 ),
+                ordenes_hist: [
+                    ...( user.ordenes_hist?.map((ord) => typeof ord === 'object' ? ord.id : '') ?? [] ),
+                    chizuOrder.id
+                ],
+            }
+        });
+
+        console.log("Orden asociada al usuario");
+        
+        const { docs: productsPayload } = await payload.find({
+            collection: 'products',
+            where:{
+                id:{
+                    in: productsIDS
+                },
+            },
+        })
+
+        productsPayload.forEach( async (product) => {
+            // const qty = products_info.find( ([id, qty]) => id === product.id )?.[1] ?? 1
+
+            await payload.update({
+                collection: 'products',
+                id: product.id,
+                data: { compras: (typeof product.compras === 'number' ? product.compras : 0) + 1 },
+            });
+
+        })
+
+        console.log("Compra registrada en la DB");
+        
+        //#endregion ~~~~~~~~~~~~~~~~~~~~~~~~~~~ Payload truco ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
         const value2 = await preference2.create({
             body: {
                 items: products,
                 back_urls: {
-                    success: "https://www.youtube.com/watch?v=TwC0Db7oerM",
-                    failure: "https://www.youtube.com/@onthecode",
-                    pending: "https://chatgpt.com",
+                    
+                    success: `${process.env.NEXT_PUBLIC_SERVER_URL}/thank-you?orderId=${chizuOrder.id}`, // crear order
+                    failure: `${process.env.NEXT_PUBLIC_SERVER_URL}/cart`,
+                    pending: `${process.env.NEXT_PUBLIC_SERVER_URL}/thank-you?orderId=${chizuOrder.id}`,
                 },
                 auto_return: "approved",
             }
         });
 
-        console.log("Preferencia creada:", value2);
+
+        console.log("Preferencia Creada: ", value2.id);
+
+
+
+        //console.log("Preferencia creada:", value2);
         return res.status(200).json({ id: value2.id });
     } catch (error) {
         console.error(error);
